Update README.md
The project is an IoT-based Vehicle Accident Detection and Notification System that uses sensors and a microcontroller to identify accidents instantly. When a sudden impact is detected, the system automatically sends an emergency alert—such as time, location, and status—to selected contacts through a Telegram bot or similar platform. The goal is to reduce response time and ensure quick help during road accidents, especially in remote or low-traffic areas.